extraced FABRID functions and moved to new files
