+++ daemon/drkey/BUILD.bazel
@@ -2,7 +2,10 @@ load("//tools/lint:go.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["client_engine.go"],
+    srcs = [
+        "client_engine.go",
+        "client_engine_fabrid.go",
+    ],
     importpath = "github.com/scionproto/scion/daemon/drkey",
     visibility = ["//visibility:public"],
     deps = [
