+++ daemon/internal/servers/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "grpc.go",
+        "grpc_fabrid.go",
         "metrics.go",
     ],
     importpath = "github.com/scionproto/scion/daemon/internal/servers",
