+++ pkg/daemon/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "apitypes.go",
         "daemon.go",
         "grpc.go",
+        "grpc_fabrid.go",
         "metrics.go",
     ],
     importpath = "github.com/scionproto/scion/pkg/daemon",
