+++ pkg/daemon/grpc.go
@@ -26,7 +26,6 @@ import (
 
 	"github.com/scionproto/scion/pkg/addr"
 	"github.com/scionproto/scion/pkg/drkey"
-	"github.com/scionproto/scion/pkg/experimental/fabrid"
 	libgrpc "github.com/scionproto/scion/pkg/grpc"
 	"github.com/scionproto/scion/pkg/private/common"
 	"github.com/scionproto/scion/pkg/private/ctrl/path_mgmt"
@@ -226,57 +225,6 @@ func (c grpcConn) DRKeyGetHostHostKey(ctx context.Context,
 	return key, nil
 }
 
-// Returns all the ASHost DRKeys for the ASes inside the meta.PathAS
-func (c grpcConn) FabridKeys(ctx context.Context, meta drkey.FabridKeysMeta,
-) (drkey.FabridKeysResponse, error) {
-
-	client := sdpb.NewDaemonServiceClient((c.conn))
-	pathASes := make([]uint64, 0, len(meta.PathASes))
-	for i := 0; i < len(meta.PathASes); i++ {
-		pathASes = append(pathASes, uint64(meta.PathASes[i]))
-	}
-	resp, err := client.FabridKeys(ctx, &sdpb.FabridKeysRequest{
-		SrcHost:  meta.SrcHost,
-		PathAses: pathASes,
-		DstAs:    uint64(meta.DstAS),
-		DstHost:  meta.DstHost,
-	})
-	if err != nil {
-		return drkey.FabridKeysResponse{}, err
-	}
-	asHostKeys := make([]drkey.FabridKey, 0, len(resp.AsHostKeys))
-	for i, key := range resp.AsHostKeys {
-		epoch := drkey.Epoch{
-			Validity: cppki.Validity{
-				NotBefore: key.EpochBegin.AsTime(),
-				NotAfter:  key.EpochEnd.AsTime(),
-			},
-		}
-		asHostKeys = append(asHostKeys, drkey.FabridKey{
-			Epoch: epoch,
-			AS:    meta.PathASes[i],
-			Key:   drkey.Key(key.Key),
-		})
-	}
-	var hostHostKey drkey.FabridKey = drkey.FabridKey{}
-	if resp.HostHostKey != nil {
-		hostHostKey = drkey.FabridKey{
-			Epoch: drkey.Epoch{
-				Validity: cppki.Validity{
-					NotBefore: resp.HostHostKey.EpochBegin.AsTime(),
-					NotAfter:  resp.HostHostKey.EpochEnd.AsTime(),
-				},
-			},
-			AS:  meta.DstAS,
-			Key: drkey.Key(resp.HostHostKey.Key),
-		}
-	}
-	return drkey.FabridKeysResponse{
-		ASHostKeys: asHostKeys,
-		PathKey:    hostHostKey,
-	}, nil
-}
-
 func (c grpcConn) Close() error {
 	return c.conn.Close()
 }
@@ -369,23 +317,6 @@ func convertPath(p *sdpb.Path, dst addr.IA) (path.Path, error) {
 		AuthLHVF: append([]byte(nil), p.EpicAuths.AuthLhvf...),
 	}
 	return res, nil
-}
-
-func fabridInfoFromPB(fi *sdpb.FabridInfo) snet.FabridInfo {
-	pbPolicies := make([]*fabrid.Policy, len(fi.Policies))
-	for i, fp := range fi.Policies {
-		pbPolicies[i] = &fabrid.Policy{
-			IsLocal:    fp.PolicyIdentifier.PolicyIsLocal,
-			Identifier: fp.PolicyIdentifier.PolicyIdentifier,
-			Index:      fabrid.PolicyID(fp.PolicyIndex),
-		}
-	}
-	return snet.FabridInfo{
-		Enabled:  fi.Enabled,
-		Policies: pbPolicies,
-		Digest:   fi.Digest,
-		Detached: fi.Detached,
-	}
 }
 
 func linkTypeFromPB(lt sdpb.LinkType) snet.LinkType {
