+++ pkg/drkey/drkey.go
@@ -199,24 +199,3 @@ type HostHostKey struct {
 	DstHost string
 	Key     Key
 }
-
-type FabridKeysMeta struct {
-	SrcAS   addr.IA
-	SrcHost string
-	// ASes on the path. Don't have to be in order
-	PathASes []addr.IA
-	// Field is optional. If DstHost is nil, no path-key will be fetched
-	DstHost *string
-	DstAS   addr.IA
-}
-
-type FabridKey struct {
-	Epoch Epoch
-	AS    addr.IA
-	Key   Key
-}
-
-type FabridKeysResponse struct {
-	ASHostKeys []FabridKey
-	PathKey    FabridKey
-}
