+++ pkg/grpc/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "creds.go",
         "dialer.go",
+        "dialer_drkey.go",
         "interceptor.go",
     ],
     importpath = "github.com/scionproto/scion/pkg/grpc",
