+++ pkg/grpc/dialer.go
@@ -50,47 +50,6 @@ func (SimpleDialer) Dial(ctx context.Context, address net.Addr) (*grpc.ClientCon
 	)
 }
 
-// The FixedLocalIPTCPDialer behaves the same as the TCP Dialer but allows for setting a local
-// IP address. Without this it would not be possible to call IP sensitive endpoints like
-// control service drkey.
-type FixedLocalIPTCPDialer struct {
-	LocalAddr   *net.TCPAddr
-	SvcResolver func(addr.SVC) []resolver.Address
-}
-
-func (d *FixedLocalIPTCPDialer) Dial(ctx context.Context, dst net.Addr) (*grpc.ClientConn, error) {
-	dialer := func(ctx context.Context, addr string) (net.Conn, error) {
-		csAddr, err := net.ResolveTCPAddr("tcp", addr)
-		if err != nil {
-			return nil, err
-		}
-		return net.DialTCP("tcp", d.LocalAddr, csAddr)
-	}
-	if v, ok := dst.(*snet.SVCAddr); ok {
-		targets := d.SvcResolver(v.SVC)
-		r := manual.NewBuilderWithScheme("svc")
-		r.InitialState(resolver.State{Addresses: targets})
-		if len(targets) == 0 {
-			return nil, serrors.New("could not resolve")
-		}
-		return grpc.DialContext(ctx, r.Scheme()+":///"+v.SVC.BaseString(),
-			grpc.WithDefaultServiceConfig(`{"loadBalancingConfig": [{"round_robin":{}}]}`),
-			grpc.WithInsecure(),
-			grpc.WithContextDialer(dialer),
-			grpc.WithResolvers(r),
-			UnaryClientInterceptor(),
-			StreamClientInterceptor(),
-		)
-	} else {
-		return grpc.DialContext(ctx, dst.String(),
-			grpc.WithInsecure(),
-			grpc.WithContextDialer(dialer),
-			UnaryClientInterceptor(),
-			StreamClientInterceptor(),
-		)
-	}
-}
-
 // XXX(roosd):
 //
 // Dialing a grpc.ClientConn is hidden behind the dialer interface due to
