+++ pkg/private/xtest/graph/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "default_gen.go",
         "graph.go",
+        "graph_fabrid.go",
     ],
     importpath = "github.com/scionproto/scion/pkg/private/xtest/graph",
     visibility = ["//visibility:public"],
