+++ pkg/snet/path/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "fabrid.go",
         "onehop.go",
         "path.go",
+        "path_ext.go",
         "scion.go",
     ],
     importpath = "github.com/scionproto/scion/pkg/snet/path",
