+++ pkg/snet/path/epic.go
@@ -94,7 +94,3 @@ func (e *EPIC) SetPath(s *slayers.SCION) error {
 	s.Path, s.PathType = ep, ep.Type()
 	return nil
 }
-
-func (e *EPIC) SetExtensions(s *slayers.SCION, p *snet.PacketInfo) error {
-	return nil
-}
