+++ proto/control_plane/experimental/v1/seg_detached_extensions_fabrid.proto
@@ -1,10 +1,10 @@
-// Copyright 2020 Anapaya Systems
+// Copyright 2024 ETH Zurich
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//   http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,9 +18,9 @@ option go_package = "github.com/scionproto/scion/pkg/proto/control_plane/experim
 
 package proto.control_plane.experimental.v1;
 
-message EPICDetachedExtension {
-    // Remaining bytes of the full MAC in the hop entry.
-    bytes auth_hop_entry = 1;
-    // Remaining bytes of the full MAC in the peer entries.
-    repeated bytes auth_peer_entries = 2;
+import "proto/control_plane/experimental/v1/fabrid_extensions.proto";
+
+message FABRIDDetachedExtension {
+    // The I and D-map of FABRID
+    proto.control_plane.experimental.v1.FABRIDDetachableMaps maps = 1;
 }
