+++ proto/control_plane/v1/seg_extensions.proto
@@ -19,6 +19,7 @@ option go_package = "github.com/scionproto/scion/pkg/proto/control_plane";
 package proto.control_plane.v1;
 
 import "proto/control_plane/experimental/v1/seg_detached_extensions.proto";
+import "proto/control_plane/experimental/v1/seg_detached_extensions_fabrid.proto";
 
 message PathSegmentExtensions {
     // Optional static info extension.
