+++ proto/daemon/v1/BUILD.bazel
@@ -4,6 +4,7 @@ proto_library(
     name = "daemon",
     srcs = [
         "daemon.proto",
+        "daemon_fabrid.proto",
     ],
     visibility = ["//visibility:public"],
     deps = [
