+++ tools/end2end/main.go
@@ -37,16 +37,11 @@ import (
 
 	"github.com/scionproto/scion/pkg/addr"
 	"github.com/scionproto/scion/pkg/daemon"
-	"github.com/scionproto/scion/pkg/drkey"
 	libfabrid "github.com/scionproto/scion/pkg/experimental/fabrid"
-	"github.com/scionproto/scion/pkg/experimental/fabrid/crypto"
 	"github.com/scionproto/scion/pkg/log"
 	"github.com/scionproto/scion/pkg/private/common"
 	"github.com/scionproto/scion/pkg/private/serrors"
 	"github.com/scionproto/scion/pkg/private/util"
-	"github.com/scionproto/scion/pkg/slayers"
-	"github.com/scionproto/scion/pkg/slayers/extension"
-	"github.com/scionproto/scion/pkg/slayers/path/scion"
 	"github.com/scionproto/scion/pkg/snet"
 	"github.com/scionproto/scion/pkg/snet/metrics"
 	snetpath "github.com/scionproto/scion/pkg/snet/path"
@@ -248,158 +243,6 @@ func (s server) handlePing(conn *snet.Conn) error {
 		return withTag(serrors.WrapStr("sending reply", err))
 	}
 	log.Info("Sent pong to", "client", clientUDPAddr)
-	return nil
-}
-
-func (s server) handlePingFabrid(conn snet.PacketConn) error {
-	var p snet.Packet
-	var ov net.UDPAddr
-	err := readFromFabrid(conn, &p, &ov)
-	if err != nil {
-		return serrors.WrapStr("reading packet", err)
-	}
-
-	// If the packet is from remote IA, validate the FABRID path
-	if p.Source.IA != integration.Local.IA {
-		if p.HbhExtension == nil {
-			return serrors.New("Missing HBH extension")
-		}
-
-		// Check extensions for relevant options
-		var identifierOption *extension.IdentifierOption
-		var fabridOption *extension.FabridOption
-		var err error
-
-		for _, opt := range p.HbhExtension.Options {
-			switch opt.OptType {
-			case slayers.OptTypeIdentifier:
-				decoded := scion.Decoded{}
-				err = decoded.DecodeFromBytes(p.Path.(snet.RawPath).Raw)
-				if err != nil {
-					return err
-				}
-				baseTimestamp := decoded.InfoFields[0].Timestamp
-				identifierOption, err = extension.ParseIdentifierOption(opt, baseTimestamp)
-				if err != nil {
-					return err
-				}
-			case slayers.OptTypeFabrid:
-				fabridOption, err = extension.ParseFabridOptionFullExtension(opt,
-					(opt.OptDataLen-4)/4)
-				if err != nil {
-					return err
-				}
-			}
-		}
-
-		if identifierOption == nil {
-			return serrors.New("Missing identifier option")
-		}
-
-		if fabridOption == nil {
-			return serrors.New("Missing FABRID option")
-		}
-
-		meta := drkey.HostHostMeta{
-			Validity: identifierOption.Timestamp,
-			SrcIA:    integration.Local.IA,
-			SrcHost:  integration.Local.Host.IP.String(),
-			DstIA:    p.Source.IA,
-			DstHost:  p.Source.Host.IP().String(),
-			ProtoId:  drkey.FABRID,
-		}
-		hostHostKey, err := integration.SDConn().DRKeyGetHostHostKey(context.Background(), meta)
-		if err != nil {
-			return serrors.WrapStr("getting host key", err)
-		}
-
-		tmpBuffer := make([]byte, (len(fabridOption.HopfieldMetadata)*3+15)&^15+16)
-		_, err = crypto.VerifyPathValidator(fabridOption, tmpBuffer, hostHostKey.Key[:])
-		if err != nil {
-			return err
-		}
-	}
-
-	udp, ok := p.Payload.(snet.UDPPayload)
-	if !ok {
-		return serrors.New("unexpected payload received",
-			"source", p.Source,
-			"destination", p.Destination,
-			"type", common.TypeOf(p.Payload),
-		)
-	}
-	var pld Ping
-	if err := json.Unmarshal(udp.Payload, &pld); err != nil {
-		return serrors.New("invalid payload contents",
-			"source", p.Source,
-			"destination", p.Destination,
-			"data", string(udp.Payload),
-		)
-	}
-
-	spanCtx, err := opentracing.GlobalTracer().Extract(
-		opentracing.Binary,
-		bytes.NewReader(pld.Trace),
-	)
-	if err != nil {
-		return serrors.WrapStr("extracting trace information", err)
-	}
-	span, _ := opentracing.StartSpanFromContext(
-		context.Background(),
-		"handle_ping",
-		ext.RPCServerOption(spanCtx),
-	)
-	defer span.Finish()
-	withTag := func(err error) error {
-		tracing.Error(span, err)
-		return err
-	}
-
-	if pld.Message != ping || !pld.Server.Equal(integration.Local.IA) {
-		return withTag(serrors.New("unexpected data in payload",
-			"source", p.Source,
-			"destination", p.Destination,
-			"data", pld,
-		))
-	}
-	log.Info(fmt.Sprintf("Ping received from %s, sending pong.", p.Source))
-	raw, err := json.Marshal(Pong{
-		Client:  p.Source.IA,
-		Server:  integration.Local.IA,
-		Message: pong,
-		Trace:   pld.Trace,
-	})
-	if err != nil {
-		return withTag(serrors.WrapStr("packing pong", err))
-	}
-
-	p.Destination, p.Source = p.Source, p.Destination
-	p.Payload = snet.UDPPayload{
-		DstPort: udp.SrcPort,
-		SrcPort: udp.DstPort,
-		Payload: raw,
-	}
-
-	// Remove header extension for reverse path
-	p.HbhExtension = nil
-	p.E2eExtension = nil
-
-	// reverse path
-	rpath, ok := p.Path.(snet.RawPath)
-	if !ok {
-		return serrors.New("unexpected path", "type", common.TypeOf(p.Path))
-	}
-	replypather := snet.DefaultReplyPather{}
-	replyPath, err := replypather.ReplyPath(rpath)
-	if err != nil {
-		return serrors.WrapStr("creating reply path", err)
-	}
-	p.Path = replyPath
-	// Send pong
-	if err := conn.WriteTo(&p, &ov); err != nil {
-		return withTag(serrors.WrapStr("sending reply", err))
-	}
-	log.Info("Sent pong to", "client", p.Destination)
 	return nil
 }
 
@@ -640,19 +483,6 @@ func readFrom(conn *snet.Conn, pld []byte) (int, net.Addr, error) {
 		return n, remoteAddr, err
 	}
 	return n, remoteAddr, serrors.WithCtx(err,
-		"isd_as", opErr.RevInfo().IA(),
-		"interface", opErr.RevInfo().IfID,
-	)
-}
-
-func readFromFabrid(conn snet.PacketConn, pkt *snet.Packet, ov *net.UDPAddr) error {
-	err := conn.ReadFrom(pkt, ov)
-	// Attach more context to error
-	var opErr *snet.OpError
-	if !(errors.As(err, &opErr) && opErr.RevInfo() != nil) {
-		return err
-	}
-	return serrors.WithCtx(err,
 		"isd_as", opErr.RevInfo().IA(),
 		"interface", opErr.RevInfo().IfID,
 	)
