+++ tools/integration/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "cmd.go",
         "docker.go",
         "done.go",
+        "endhost_integration.go",
         "integration.go",
         "networkalloc.go",
     ],
